# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WASAtext
  description: |-
    homework wasa
  version: 0.0.1
...
paths:
...
  /profiles/{user_id}/conversations/{dest_user_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/convo_id"
    put: 
      tags: ["Conversation"]
      summary: create new conversation
      operationId: createConversation
      description: create new conversation, unless it doesn't already exists
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema: {$ref: "#/components/schemas/NewUserId"}
        required: true
      responses:
        "200":
          description: convo created succesfully
          content:
           application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }     
    get:
      tags: ["Conversation"]
      summary: open existing conversation
      operationId: getConversation
      description: open an already created conversation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: convo opened succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Conversation" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      tags: ["Conversation"]
      summary: delete existing conversation
      operationId: deleteConversation
      description: user deletes conversation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: convo opened succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
...
  /profiles/{user_id}/conversations/{dest_user_id}/messages/{msg_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/convo_id"
      - $ref: "#/components/parameters/msg_id"
    delete:
      tags: ["Conversation"]
      summary: delete message
      operationId: deleteMessage
      description: user deletes message
      security:
        - bearerAuth: []
      responses:
        "200":
          description: message deleted succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetConvoID:
                  type: integer
                  description: ID della conversazione con cui `userID` vuole condividere il messaggio inoltrato
                customContent:
                  type: string
                  description: (Opzionale) Testo aggiuntivo che l’utente può scrivere accanto al messaggio inoltrato
      tags: ["Conversation"]
      summary: forward message
      description: send existing message in the selected conversation
      operationId: forwardMessage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: message sent succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profiles/{user_id}/groups:
    parameters:
      - $ref: "#/components/parameters/user_id"
    post:
      tags: ["Group"]
      summary: create group
      operationId: createGroup
      description: user creates group
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              description: user inputs name and partecipants of the group
              type: object
              properties:
                name: { $ref: "#/components/schemas/GroupName"}
                users: { $ref: "#/components/schemas/ProfilesList"}
        required: true
      responses:
        "200":
          description: group created succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Group" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    get:
      summary: list all groups
      tags: ["Group"]
      operationId: getMyGroups
      description: Obtain list of all user's groups, in reverse order
      security:
        - bearerAuth: []
      responses:
        "200":
          description: groups list obtained succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/GroupsList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{user_id}/groups/{group_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/group_id"
    get:
      tags: ["Group"]
      summary: open existing group
      operationId: getGroup
      description: open an already created group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: group opened succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Group" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      tags: ["Group"]
      summary: user leaves group
      operationId: leaveGroup
      description: user can leave given group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: group left succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{user_id}/groups/{group_id}/messages:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/group_id"
    get:
      summary: list all messages in a group
      tags: ["Group"]
      operationId: getGroupMessages
      description: Obtain list of all messages in a group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: message list obtained succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/MessagesList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      parameters:
        - name: reply
          in: query
          required: false
          schema: {$ref: "#/components/schemas/Identificator"}
      tags: ["Group"]
      summary: send message
      operationId: sendGroupMessage
      description: send new message in the selected group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/Message"}
      responses:
        "200":
          description: message sent succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{user_id}/groups/{group_id}/messages/{msg_id}:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/group_id"
      - $ref: "#/components/parameters/msg_id"
    delete:
      tags: ["Group"]
      summary: delete message
      operationId: deleteGroupMessage
      description: user deletes message
      security:
        - bearerAuth: []
      responses:
        "200":
          description: message deleted succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                targetGroupID:
                  type: integer
                  description: ID del gruppo con cui `userID` vuole condividere il messaggio inoltrato
                customContent:
                  type: string
                  description: (Opzionale) Testo aggiuntivo che l’utente può scrivere accanto al messaggio inoltrato
      tags: ["Group"]
      summary: forward message
      description: send existing message in the selected group
      operationId: forwardGroupMessage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: message sent succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profiles/{user_id}/groups/{group_id}/members:
    parameters:
      - $ref: "#/components/parameters/user_id"
      - $ref: "#/components/parameters/group_id"
    put:
      tags: ["Group"]
      summary: add specific user to group
      operationId: addToGroup
      description: user can add a new user to a specific group
      security:
        - bearerAuth: []
      requestBody:
        description: user to add
        content:
          application/json:
            schema:
              type: object
              properties:
                new_user_id:
                  $ref: "#/components/schemas/Identificator"
        required: true
      responses:
        "200":
          description: user added succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    get:
      summary: list all members in a group
      tags: ["Group"]
      operationId: getGroupMembers
      description: Obtain list of all members in a group
      security:
        - bearerAuth: []
      responses:
        "200":
          description: member list obtained succesfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/ProfilesList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }

#---------------------------COMPONENTS--------------------------#
components:
  parameters: #tutto cio che prendi nei path
    user_id:
      name: user_id
      description: id of a given user
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identificator" }
    convo_id:
      name: dest_user_id
      description: id of a conversation user
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identificator" }
    msg_id:
      name: msg_id
      description: id of a given message
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identificator" }
...
    group_id:
      name: group_id
      description: id of a given group
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Identificator" }
  schemas:
    Identificator:
      description: id for user/conversation/group/message
      type: integer
      minimum: 1
      maximum: 999
      example: 9
    Username:
      description: name of user
      type: string
      pattern: ^[a-z0-9]{3,13}$
      example: "octopus24"
      minLength: 3
      maxLength: 13
    NewUserId:
      description: id of another user
      type: object
      properties:  
        id: {$ref: "#/components/schemas/Identificator"}
      required:
        - id 
    GroupName:
      description: name of group
      type: string
      pattern: ^[A-Za-zÀ-ÖØ-öø-ÿ0-9 ]{3,100}(/[A-Za-z])?$
      example: "Informatica Canale M/Z"
      minLength: 3 
      maxLength: 100
    PhotoOrGif:
      description: image
      type: object
      properties:
        photo:
          description: image or gif
          type: string
          format: byte
          minLength: 0
          maxLength: 9999
    Profile:
      type: object
      description: user profile
      properties:
        user_id: { $ref: "#/components/schemas/Identificator" }
        username: { $ref: "#/components/schemas/Username" }
        photo: { $ref: "#/components/schemas/PhotoOrGif" }
    ProfilesList:
      type: object
      description: searched user's list
      properties:
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
          minItems: 0 
          maxItems: 999
      required:
        - profiles 
...
    Conversation:
      description: user's conversation
      type: object
      properties:
        convo_id: { $ref: "#/components/schemas/Identificator" }
        user_id: { $ref: "#/components/schemas/Identificator" }
    GroupsList:
      type: object
      description: user's groups list       
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Group"
          minItems: 0 
          maxItems: 999
      required:
        - groups
    Group:
      description: group 
      type: object 
      properties:
        g_id: { $ref: "#/components/schemas/Identificator"}
        g_name: { $ref: "#/components/schemas/GroupName"}
        g_photo: { $ref: "#/components/schemas/PhotoOrGif"}
        users: { $ref: "#/components/schemas/ProfilesList"}
      required:
        - g_name
        - g_id
        - users 
    Message:
      description: user's message in a convo/group
      type: object
      properties:
        msg_id: { $ref: "#/components/schemas/Identificator" }
        timestamp: { $ref: "#/components/schemas/Timestamp" }
        content:
          oneOf:
            - $ref: "#/components/schemas/Text"
            - $ref: "#/components/schemas/PhotoOrGif"
      required: 
        - msg_id
        - timestamp
    MessagesList:
      type: object
      description: list of messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
          minItems: 0 
          maxItems: 999
      required:
        - messages 
...
    Text:
      type: object
      description: text of the message
      properties:
        text:
          type: string
          example: "Hey, come stai?"
          pattern: '^[^\/\\]{0,64}$'
          minLength: 1
          maxLength: 65536
...
    Timestamp:
      title: timestamp
      description: the timestamp of the message
      type: string
      format: date-time
      pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
      example: "2020-12-31T23:59:59"
      minLength: 19
      maxLength: 19
    Ok:
      title: OK
      description: The succesfull response from server.
      type: string
      example: "OK, action successfully completed"
      minLength: 2
      maxLength: 200
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
  responses:
...
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        text/plain:
          schema:
            description: BadRequest
            type: string
            example: "Bad Request"
            minLength: 11
            maxLength: 30

    #401
    Unauthorized:
      description: The access token is missing or it's expired.
      content:
        text/plain:
          schema:
            description: Unauthorized
            type: string
            example: "Unauthorized"
            minLength: 12
            maxLength: 30

    #403
    Forbidden:
      description: The user is not allowed to perform the action.
      content:
        text/plain:
          schema:
            description: Forbidden
            type: string
            example: "Forbidden"
            minLength: 9
            maxLength: 30

    #500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        text/plain:
          schema:
            description: InternalServerError
            type: string
            example: "Internal Server Error"
            minLength: 21
            maxLength: 30